feat: replace iTunes link (#2448)
* feat: replace iTunes link

* fix: handle apps with empty update logs
