+++ lib/router.js
@@ -468,11 +468,11 @@ router.get('/pigtails', require('./routes/pigtails'));
 router.get('/imuseum/:city/:type?', require('./routes/imuseum'));
 
 // AppStore
-router.get('/appstore/update/:country/:id', require('./routes/appstore/update'));
-router.get('/appstore/price/:country/:type/:id', require('./routes/appstore/price'));
-router.get('/appstore/iap/:country/:id', require('./routes/appstore/in-app-purchase'));
-router.get('/appstore/xianmian', require('./routes/appstore/xianmian'));
-router.get('/appstore/gofans', require('./routes/appstore/gofans'));
+router.get('/appstore/update/:country/:id', require('./routes/apple/appstore/update'));
+router.get('/appstore/price/:country/:type/:id', require('./routes/apple/appstore/price'));
+router.get('/appstore/iap/:country/:id', require('./routes/apple/appstore/in-app-purchase'));
+router.get('/appstore/xianmian', require('./routes/apple/appstore/xianmian'));
+router.get('/appstore/gofans', require('./routes/apple/appstore/gofans'));
 
 // Hopper
 router.get('/hopper/:lowestOnly/:from/:to?', require('./routes/hopper/index'));
