+++ lib/routes/apple/appstore/in-app-purchase.js
@@ -5,7 +5,7 @@ const cheerio = require('cheerio');
 module.exports = async (ctx) => {
     const country = ctx.params.country;
     const id = ctx.params.id;
-    const link = `https://itunes.apple.com/${country}/app/${id}`;
+    const link = `https://apps.apple.com/${country}/app/${id}`;
     const target = url.resolve(link, '?mt=8#see-all/in-app-purchases');
 
     const res = await got.get(target);
