+++ lib/routes/apple/appstore/price.js
@@ -33,7 +33,7 @@ module.exports = async (ctx) => {
 
     const title = `${country === 'cn' ? '限免提醒' : 'Price watcher'}: ${result.title} for ${type === 'macapps' ? 'macOS' : 'iOS'}`;
 
-    const link = `https://itunes.apple.com/${country}/app/id${id}`;
+    const link = `https://apps.apple.com/${country}/app/id${id}`;
 
     if (result.priceDropIndicator === -1) {
         const single = {
